refactor: Move Cancellation to ride package
