+++ Shuttle-back/src/main/java/com/shuttle/ride/Cancellation.java
@@ -1,6 +1,5 @@
-package com.shuttle.panic;
+package com.shuttle.ride;
 
-import com.shuttle.ride.Ride;
 import com.shuttle.user.GenericUser;
 import jakarta.persistence.*;
 import lombok.Data;
@@ -10,15 +9,13 @@ import java.time.LocalDateTime;
 @Data
 @Entity
 public class Cancellation {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     @ManyToOne
     private GenericUser user;
-    @ManyToOne // TODO: Should be OneToOne
+    @ManyToOne // TODO: Should be OneToOne?
     private Ride ride;
     private LocalDateTime time;
     private String reason;
-
 }
